+++ app/Facades/Schedule.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Facades;
+
+use Illuminate\Support\Facades\Facade;
+
+class Schedule extends Facade
+{
+    protected static function getFacadeAccessor()
+    {
+        return 'ScheduleService';
+    }
+}
