+++ app/Providers/AppServiceProvider.php
@@ -3,6 +3,7 @@
 namespace App\Providers;
 
 use App\Contracts\UploadImageServiceContract;
+use App\Services\ScheduleService;
 use App\Services\ImageUploadService;
 use Illuminate\Support\ServiceProvider;
 use Illuminate\Support\Facades\Schema;
@@ -17,6 +18,7 @@ class AppServiceProvider extends ServiceProvider
     public function register()
     {
         $this->app->bind(UploadImageServiceContract::class, ImageUploadService::class);
+        $this->app->bind('ScheduleService', ScheduleService::class);
     }
 
     /**
